% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lua.R
\name{lua}
\alias{lua}
\title{Run Lua code}
\usage{
lua(code, filename = NULL, L = NULL)
}
\arguments{
\item{code}{Lua code block to run.}

\item{filename}{If non-\code{NULL}, name of file to run.}

\item{L}{\link[=lua_open]{Lua state} in which to run the code. \code{NULL} (default)
uses the default Lua state for \pkg{luajr}.}
}
\value{
Lua value(s) returned by the code block converted to R object(s).
Only a subset of all Lua types can be converted to R objects at present.
If multiple values are returned, these are packaged in a \code{list}.
}
\description{
Runs the specified Lua code.
}
\examples{
twelve <- lua("return 3*4")
print(twelve)
}
