% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lua_reset}
\alias{lua_reset}
\title{Reset the default Lua state}
\usage{
lua_reset()
}
\description{
Clears out all variables from the default Lua state, freeing up the
associated memory.
}
\details{
To reset a Lua state returned by \code{\link[=lua_open]{lua_open()}}, just call \code{\link[=lua_open]{lua_open()}} again.
The memory previously used will be cleaned up at the next garbage
collection.
}
\examples{
lua("a = 2")
lua_reset()
lua("print(a)") # nil
}
