% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lua_shell.R
\name{lua_shell}
\alias{lua_shell}
\title{Run an interactive Lua shell}
\usage{
lua_shell(L = NULL)
}
\arguments{
\item{L}{\link[=lua_open]{Lua state} in which to run the code. \code{NULL} (default)
uses the default Lua state for \pkg{luajr}.}
}
\value{
None.
}
\description{
When in interactive mode, provides a basic read-eval-print loop with LuaJIT.
}
\details{
As a convenience, lines starting with an equals sign have the \code{"="} replaced
with \code{"return "}, so that e.g. entering \verb{=x} will show the value of \code{x} as
returned to R.
}
