% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luajr-package.R
\docType{package}
\name{luajr}
\alias{luajr}
\title{luajr: Use LuaJIT Scripting with R}
\description{
'luajr' provides an interface to Mike Pall's LuaJIT (<https://luajit.org>),
a just-in-time compiler for the Lua scripting language
(<https://www.lua.org>). It allows users to run Lua code from R.
}
\section{The R API}{

* [lua()]
* [lua_func()]
* [lua_shell()]
* [lua_open()]
}

\section{Other information}{

* [using luajr from C++]
* [using luajr from Lua]

For Introduction, see \code{vignette("intro", package = "luajr")}
}

