% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lua_func.R
\name{lua_func}
\alias{lua_func}
\title{Make a Lua function callable from R}
\usage{
lua_func(func, argcode = ".", L = NULL)
}
\arguments{
\item{func}{A character string with a Lua expression evaluating to a
function, or an external pointer to a Lua function.}

\item{argcode}{How to wrap R arguments for the Lua function.}

\item{L}{\link[=lua_open]{Lua state} in which to run the code. \code{NULL} (default)
uses the default Lua state for \pkg{luajr}.}
}
\value{
An R function which can be called to invoke the Lua function.
}
\description{
Takes any Lua expression (as a character string) that evaluates to a
function and provides an R function that can be called to invoke the Lua
function. Instead of a character string, you can also provide an external
pointer to a Lua function (see examples).
}
\details{
The R types that can be passed to Lua are: \code{NULL}, logical vector,
integer vector, numeric vector, string vector, list, external pointer, and
raw.

The parameter \code{argcode} is a string with one character for each argument of
the Lua function, recycled as needed (e.g. so that a single character would
apply to all arguments regardless of how many there are).

In the following, the corresponding character of \code{argcode} for a specific
argument is referred to as its arg code.

For \code{NULL} or any argument with length 0, the result in Lua is \strong{nil}
regardless of the corresponding arg code.

For logical, integer, double, and character vectors, if the corresponding
arg code is \code{'s'} (simplify), then if the R vector has length one, it is
supplied as a Lua primitive (boolean, number, number, or string,
respectively), and if length > 1, as an array, i.e. a table with integer
indices starting at 1. If the code is \code{'a'}, the vector is always supplied as
an array, even if it only has length 1. If the arg code is the digit \code{'1'}
through \code{'9'}, this is the same as \code{'s'}, but the vector is required to have
that specific length, otherwise an error message is emitted.

Still focusing on the same vector types, if the arg code is \code{'r'}, then the
vector is passed \emph{by reference} to Lua, adopting the type \code{luajr.logical},
\code{luajr.integer}, \code{luajr.numeric}, or \code{luajr.character} as appropriate.
Modifications to individual elements will be reflected in the R vector.
If the arg code is \code{'v'}, the vector is passed \emph{by value} (copy-on-write)
to Lua, adopting the same types. A copy is made on first modification.

For a raw vector, only the \code{'s'} type is accepted and the result in Lua is
a string (potentially with embedded nulls).

For lists, if the arg code is \code{'s'} (simplify), the list is passed as a Lua
table. Any entries of the list with non-blank names are named in the table,
while unnamed entries have the associated integer key in the table. Note that
Lua does not preserve the order of entries in tables. This means that an R
list with names will often go "out of order" when passed into Lua with \code{'s'}
and then returned back to R. This is avoided with arg code \code{'r'} or \code{'v'}.

If a list is passed in with the arg code \code{'r'} or \code{'v'}, the list is
passed to Lua as type \code{luajr.list}, and all vector elements of the list are
passed by reference or by value, respectively.

If the arg code is \code{'x'}, the value is passed as a bare SEXP (i.e. the raw R
object pointer). This bypasses all type-specific conversions and is intended
for use with the \code{R} Lua module (loaded via \code{require("R")}).

For external pointers, the arg code is ignored and the external pointer is
passed to Lua as type \strong{userdata}.

When the function is called and Lua values are returned from the function,
the Lua return values are converted to R values as follows.

If nothing is returned, the function returns \code{invisible()} (i.e. \code{NULL}).

If multiple arguments are returned, a list with all arguments is returned.

Vector types (e.g. \code{luajr.logical}) are returned to R as the corresponding R
vector. A \code{luajr.list} is returned as an R list. Vector and list types
respect R attributes set within Lua code.

A \strong{table} is returned as a list. In the list, any table entries with a
number key come first (with indices 1 to n, i.e. the original number key's
value is discarded), followed by any table entries with a string key
(named accordingly). This may well scramble the order of keys, so beware.
Note in particular that Lua does not guarantee that it will traverse a table
in ascending order of keys. Entries with non-number, non-string keys are
discarded. It is probably best to avoid returning a \strong{table} with anything
other than string keys, or to use \code{luajr.list}.

A Lua string with embedded nulls is returned as an R raw type.

A function is returned as an external pointer, which itself can be converted
into a function that can be called from R, by passing it through \code{lua_func()}
as the \code{func} argument.
}
\examples{
# use with a character string
squared <- lua_func("function(x) return x^2 end")
print(squared(7))

# use with an external pointer to a Lua function
times2ptr <- lua("return function(x) return 2 * x end")
print(times2ptr)
times2 <- lua_func(times2ptr)
print(times2(14))
}
