% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lua_func.R
\name{lua_func}
\alias{lua_func}
\title{Make a Lua function callable from R}
\usage{
lua_func(func, argcode = ".", L = NULL)
}
\arguments{
\item{func}{A character string with a Lua expression evaluating to a
function, or an external pointer to a Lua function.}

\item{argcode}{How to wrap R arguments for the Lua function.}

\item{L}{\link[=lua_open]{Lua state} in which to run the code. \code{NULL} (default)
uses the default Lua state for \pkg{luajr}.}
}
\value{
An R function which can be called to invoke the Lua function.
}
\description{
Takes any Lua expression (as a character string) that evaluates to a
Lua function and provides an R function that can be called to invoke the Lua
function. Instead of a character string, you can also provide an external
pointer to a Lua function (see examples).
}
\details{
Any R type can be passed to a Lua function. The R types that have special
support in \pkg{luajr} are: \code{NULL}, logical vector, integer vector, numeric
vector, character vector, list, function, environment, and external pointer.

The parameter \code{argcode} is a string that specifies type handling for each
argument of the Lua function. The last type is repeated when there are more
arguments than specified types.

To receive a \code{luajr.logical} type in Lua, use the argcode \code{logical} or \code{L}.
Similarly, \code{luajr.integer} is specified with \code{integer} or \code{I};
\code{luajr.numeric} is specified with \code{numeric} or \code{N}; and \code{luajr.character} is
specified with \code{character} or \code{C}. \code{luajr.list} is specified with \code{list},
\code{vector}, or \code{V}.

To receive a \code{luajr.rfunction} type, use \code{rfunction} or \code{F}. To receive a
\code{luajr.environment} type, use \code{environment} or \code{E}. To receive an \code{R.sexp}
type, use \code{sexp} or \code{S}.

To pass any R type as "closest available" type from the above, use the
argcode \code{auto} or \code{"."}.

Argcodes should be separated by commas. If you are using a "single-character"
argcode like \code{.} or \code{S}, there does not need to be a comma following it.

You can also receive R values as Lua native types. Specifically, the argcode
\code{function} corresponds to a Lua function; \code{boolean} is a Lua boolean;
\code{number} is a Lua number; \code{string} is a Lua string and \code{table} is a Lua
table. \code{pointer} or \code{P} passes an external pointer (EXTPTRSXP) as a Lua
light userdata. The argcode \code{native} selects the closest available native
type.

Additionally, the prefix \code{$} specifies that the type should be passed as a
Lua native type. So \verb{$auto} is equivalent to \code{native} and \verb{$C} is equivalent
to \code{string}.

The prefix \code{!} specifies strict type handling (i.e., error if the type cannot
be converted). \code{!} also yields an error on passing \code{NULL} if the argcode
specifies a Lua native type (otherwise, \code{NULL} gets passed as Lua \code{nil}).

The prefix \code{&} can be used with logical, integer, numeric, character, or
generic (list) vectors and specifies that the value should be passed
\strong{by reference}. Vectors passed by reference can have their elements
mutated in Lua code, but they cannot be resized.

When the function is called and Lua values are returned from the function,
the Lua return values are converted to R values as follows.

If nothing is returned, the function returns \code{invisible()} (i.e. \code{NULL}).

If multiple arguments are returned, a list with all arguments is returned.

Vector types (e.g. \code{luajr.logical}) are returned to R as the corresponding R
vector. A \code{luajr.list} is returned as an R list. Vector and list types
respect R attributes set within Lua code.

A \strong{table} is returned as a list. In the list, any table entries with a
number key come first (with indices 1 to n, i.e. the original number key's
value is discarded), followed by any table entries with a string key
(named accordingly). This may well scramble the order of keys, so beware.
Note in particular that Lua does not guarantee that it will traverse a table
in ascending order of keys. Entries with non-number, non-string keys are
discarded. It is probably best to avoid returning a \strong{table} with anything
other than string keys, or to use \code{luajr.list}.

A Lua string with embedded nulls is returned as an R raw type.

A function is returned as an external pointer, which itself can be converted
into a function that can be called from R, by passing it through \code{lua_func()}
as the \code{func} argument.
}
\examples{
# use with a character string
squared <- lua_func("function(x) return x^2 end", "$.")
print(squared(7))

# use with an external pointer to a Lua function
times2ptr <- lua("return function(x) return 2 * x end")
print(times2ptr)
times2 <- lua_func(times2ptr, "$.")
print(times2(14))
}
