% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lua_shell.R
\name{lua_shell}
\alias{lua_shell}
\title{Run an interactive LuaJIT shell}
\usage{
lua_shell(L = NULL)
}
\arguments{
\item{L}{\link[=lua_open]{Lua state} in which to run the code. \code{NULL} (default)
uses the default Lua state for \pkg{luajr}.}
}
\value{
Nothing.
}
\description{
When in interactive mode, provides a basic read-eval-print loop with LuaJIT.
}
