% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luajr-package.R
\docType{package}
\name{luajr-package}
\alias{luajr}
\alias{luajr-package}
\title{luajr: LuaJIT Scripting}
\description{
'luajr' provides an interface to Mike Pall's LuaJIT (\url{https://luajit.org}),
a just-in-time compiler for the Lua scripting language
(\url{https://www.lua.org}). It allows users to run Lua code from R.
}
\section{The R API}{

\itemize{
\item \code{\link[=lua]{lua()}}: run Lua code
\item \code{\link[=lua_func]{lua_func()}}: make a Lua function callable from R
\item \code{\link[=lua_shell]{lua_shell()}}: run an interactive Lua shell
\item \code{\link[=lua_open]{lua_open()}}: create a new Lua state
\item \code{\link[=lua_reset]{lua_reset()}}: reset the default Lua state
}
}

\section{Further reading}{

For an introduction to 'luajr', see \code{vignette("luajr")}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nicholasdavies/luajr}
  \item \url{https://nicholasdavies.github.io/luajr/}
  \item Report bugs at \url{https://github.com/nicholasdavies/luajr/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Davies \email{nicholas.davies@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-1740-1412}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Mike Pall (Author of the embedded LuaJIT compiler) [copyright holder]
  \item Lua.org, PUC-Rio (Authors of the Lua scripting language) [copyright holder]
}

}
