% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luajr-package.R
\docType{package}
\name{luajr}
\alias{luajr}
\title{luajr: LuaJIT Scripting}
\description{
'luajr' provides an interface to Mike Pall's LuaJIT (\url{https://luajit.org}),
a just-in-time compiler for the Lua scripting language
(\url{https://www.lua.org}). It allows users to run Lua code from R.
}
\section{The R API}{

\itemize{
\item \code{\link[=lua]{lua()}}: run Lua code
\item \code{\link[=lua_func]{lua_func()}}: make a Lua function callable from R
\item \code{\link[=lua_shell]{lua_shell()}}: run an interactive Lua shell
\item \code{\link[=lua_open]{lua_open()}}: create a new Lua state
\item \code{\link[=lua_reset]{lua_reset()}}: reset the default Lua state
}
}

\section{Further reading}{

For an introduction to 'luajr', see \code{vignette("luajr")}
}

