% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luajr-package.R
\docType{package}
\name{luajr}
\alias{luajr}
\title{luajr: LuaJIT Scripting}
\description{
'luajr' provides an interface to Mike Pall's LuaJIT (\url{https://luajit.org}),
a just-in-time compiler for the Lua scripting language
(\url{https://www.lua.org}). It allows users to run Lua code from R.
}
\section{The R API}{

\itemize{
\item \code{\link[=lua]{lua()}}
\item \code{\link[=lua_func]{lua_func()}}
\item \code{\link[=lua_open]{lua_open()}}
\item \code{\link[=lua_reset]{lua_reset()}}
\item \code{\link[=lua_shell]{lua_shell()}}
}
}

\section{Other information}{

\itemize{
\item \link{using luajr from C++}
\item \link{using luajr from Lua}
}

For Introduction, see \code{vignette("intro", package = "luajr")}
}

